Call `entry_url_from_entry.to_s` in conditional
`entry_url_from_entry` returns an object, which is always evaluated
in a truthy way. In a rare case, this would result in that method's
`to_s` being called which would return `nil` which would then be
sent to `absolutize`.

So…

`entry_url_from_entry || @source` => <Microformats2::Property::Url>

That, converted to a string => `nil`

absolutize(nil) => 💣 URI::InvalidURIError: bad URI(is not URI?)